+++ homework/grayscale-image/compression.hpp
@@ -9,4 +9,3 @@ constexpr size_t height{32};
 
 std::vector<std::pair<uint8_t, uint8_t>> compressGrayscale(const std::array<std::array<uint8_t, width>, height>& arr);
 std::array<std::array<uint8_t, width>, height> decompressGrayscale(const std::vector<std::pair<uint8_t, uint8_t>>& vec);
-void printMap(const std::array<std::array<uint8_t, width>, height>& arr);
